+++ packages/matter.js/src/cluster/server/GroupKeyManagementServer.ts
@@ -26,7 +26,7 @@ export const GroupKeyManagementClusterHandler: () => ClusterServerHandlers<typeo
 
     keySetWrite: async () => {
         throw new StatusResponseError(
-            "We do not support and additional groups then the IPK",
+            "We do not support additional groups beyond the IPK",
             StatusCode.ResourceExhausted,
         );
     },