+++ packages/protocol/src/interaction/index.ts
@@ -8,7 +8,6 @@ export * from "./AccessControlManager.js";
 export * from "./AttributeDataDecoder.js";
 export * from "./AttributeDataEncoder.js";
 export * from "./EventDataDecoder.js";
-export * from "./EventHandler.js";
 export * from "./InteractionClient.js";
 export * from "./InteractionEndpointStructure.js";
 export * from "./InteractionMessenger.js";